allow public as documentroot to expose less to the world
The public folder is now dynamically set depending on the index.php
used. By allowing to run the application with public as document root
you can reduce the attac surface drastically, and you also reduce the
configuration overhead to protect sensitive files via webserver
configuration.

The artisan command was also "duplicated" so the AKAUNTING_PUBLIC
constant is defined in all cases.

Currently you still have to expose vendor to the world in public because
of the direct references to the adminlte package's static files. This
will be updated in a following commit.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>