Update .htaccess
I have experienced the issue where during OTA update wizard, the system files has been updated along with the .htaccess, then my directory become 403 forbidden because of disabling of visit to the front controller directory. This causes my website is not able to visit, and the update is not success. Please add the allow access code to the publoic directory. Thanks.